Camera and render engine fixes

+++ src/main/java/com/churkovainteam/kghomework/math/Point2f.java
@@ -1,13 +1,13 @@
 package com.churkovainteam.kghomework.math;
 
 public final class Point2f {
-    public double x;
-    public double y;
+    public float x;
+    public float y;
 
     public Point2f() {
     }
 
-    public Point2f(double x, double y) {
+    public Point2f(float x, float y) {
         this.x = x;
         this.y = y;
     }
