+++ src/main/java/com/churkovainteam/kghomework/math/Vector3f.java
@@ -4,14 +4,14 @@ package com.churkovainteam.kghomework.math;
 import java.util.Objects;
 
 public final class Vector3f {
-    public double x;
-    public double y;
-    public double z;
+    public float x;
+    public float y;
+    public float z;
 
     public Vector3f() {
     }
 
-    public Vector3f(double x, double y, double z) {
+    public Vector3f(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
@@ -35,8 +35,8 @@ public final class Vector3f {
             throw new IllegalArgumentException("Vector3f can not be null");
         }
 
-        double x = v1.y * v2.z - v1.z * v2.y;
-        double y = v2.x * v1.z - v2.z * v1.x;
+        float x = v1.y * v2.z - v1.z * v2.y;
+        float y = v2.x * v1.z - v2.z * v1.x;
 
         this.z = v1.x * v2.y - v1.y * v2.x;
         this.x = x;
@@ -44,15 +44,15 @@ public final class Vector3f {
     }
 
     public void normalize() {
-        double norm =
-                (1.0 / Math.sqrt(this.x * this.x + this.y * this.y + this.z * this.z));
+        float norm =
+                (float) (1.0 / Math.sqrt(this.x * this.x + this.y * this.y + this.z * this.z));
 
         this.x *= norm;
         this.y *= norm;
         this.z *= norm;
     }
 
-    public double dot(Vector3f v1) {
+    public float dot(Vector3f v1) {
         if (v1 == null) {
             throw new IllegalArgumentException("Vector3f can not be null");
         }
@@ -72,42 +72,42 @@ public final class Vector3f {
     }
 
     public void rotateAroundX(float angle) {
-        double y = (this.y * Math.cos(angle) + this.z * Math.sin(angle));
-        double z = (-this.y * Math.sin(angle) + this.z * Math.cos(angle));
+        float y = (float) (this.y * Math.cos(angle) + this.z * Math.sin(angle));
+        float z = (float) (-this.y * Math.sin(angle) + this.z * Math.cos(angle));
 
         this.y = y;
         this.z = z;
     }
 
-    public void rotateAroundY(double angle) {
-        double x = (this.x * Math.cos(angle) + this.z * Math.sin(angle));
-        double z = (-this.x * Math.sin(angle) + this.z * Math.cos(angle));
+    public void rotateAroundY(float angle) {
+        float x = (float) (this.x * Math.cos(angle) + this.z * Math.sin(angle));
+        float z = (float) (-this.x * Math.sin(angle) + this.z * Math.cos(angle));
 
         this.x = x;
         this.z = z;
     }
 
     public void rotateAroundZ(float angle) {
-        double x = (this.x * Math.cos(angle) + this.y * Math.sin(angle));
-        double y = (-this.x * Math.sin(angle) + this.y * Math.cos(angle));
+        float x = (float) (this.x * Math.cos(angle) + this.y * Math.sin(angle));
+        float y = (float) (-this.x * Math.sin(angle) + this.y * Math.cos(angle));
 
         this.x = x;
         this.y = y;
     }
 
-    public void scaleX(double scale) {
+    public void scaleX(float scale) {
         this.x *= scale;
     }
 
-    public void scaleY(double scale) {
+    public void scaleY(float scale) {
         this.y *= scale;
     }
 
-    public void scaleZ(double scale) {
+    public void scaleZ(float scale) {
         this.z *= scale;
     }
 
-    public static double angleBetweenTwoVectors(Vector3f vec1, Vector3f vec2) {
+    public static float angleBetweenTwoVectors(Vector3f vec1, Vector3f vec2) {
         if (vec1 == null || vec2 == null) {
             throw new IllegalArgumentException("Vector can't be null");
         }
@@ -116,8 +116,8 @@ public final class Vector3f {
                 (vec1.length() * vec2.length());
     }
 
-    public double length() {
-        return Math.sqrt(this.x * this.x + this.y * this.y + this.z * this.z);
+    public float length() {
+        return (float) Math.sqrt(this.x * this.x + this.y * this.y + this.z * this.z);
     }
 
     @Override
@@ -130,7 +130,9 @@ public final class Vector3f {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Vector3f vector3f = (Vector3f) o;
-        return Double.compare(vector3f.x, x) == 0 && Double.compare(vector3f.y, y) == 0 && Double.compare(vector3f.z, z) == 0;
+        return Float.compare(vector3f.x, x) == 0 &&
+                Float.compare(vector3f.y, y) == 0 &&
+                Float.compare(vector3f.z, z) == 0;
     }
 
     @Override
